+++ spec/server/brisketSpec.js
@@ -3,28 +3,12 @@
 describe("public interface to Brisket", function() {
     var Brisket = require("../../lib/brisket");
 
-    it("exposes Brisket.RouterBrewery", function() {
-        expect(Brisket.RouterBrewery).toBe(requireFromLib("controlling/RouterBrewery"));
-    });
-
-    it("exposes Brisket.Routers", function() {
-        expect(Brisket.Routers).toBe(requireFromLib("controlling/Routers"));
+    it("exposes Brisket.App", function() {
+        expect(Brisket.App).toBe(requireFromLib("application/App"));
     });
 
     it("exposes Brisket.Router", function() {
         expect(Brisket.Router).toBe(requireFromLib("controlling/Router"));
-    });
-
-    it("exposes Brisket.Controller", function() {
-        expect(Brisket.Controller).toBe(requireFromLib("controlling/Controller"));
-    });
-
-    it("exposes Brisket.Model", function() {
-        expect(Brisket.Model).toBe(requireFromLib("modeling/Model"));
-    });
-
-    it("exposes Brisket.Collection", function() {
-        expect(Brisket.Collection).toBe(requireFromLib("modeling/Collection"));
     });
 
     it("exposes Brisket.View", function() {
@@ -49,14 +33,6 @@ describe("public interface to Brisket", function() {
 
     it("exposes Brisket.Templating.StringTemplateAdapter", function() {
         expect(Brisket.Templating.StringTemplateAdapter).toBe(requireFromLib("templating/StringTemplateAdapter"));
-    });
-
-    it("exposes Brisket.ServerApp", function() {
-        expect(Brisket.ServerApp).toBe(requireFromLib("server/ServerApp"));
-    });
-
-    it("exposes Brisket.ClientApp", function() {
-        expect(Brisket.ClientApp).toBe(requireFromLib("client/ClientApp"));
     });
 
     it("exposes Brisket.Testing", function() {